Fix Negotiate auth on non-domain-joined Windows hosts
It recently came to my attention that the current use of requests-negotiate-sspi on Windows hosts was causing the module to fail if Negotiate authentication failed. This was due to our user-agent spoofing causing ADFS to continue redirecting us through the Integrated Windows Authentication (IWA) page's 401 response, since it only supports Negotiate auth - not form input.

This change passes the username and password through to the auth module, which lets Windows handle authentication properly in cases where Kerberos credentials are not available or are not accepted by the ADFS host.