Fix for DUO Universal Prompt authentication
Changing re.sub in line 147 of this file to look for ANY digit in front of the "v" part. We ran into an issue in our environment because the original code here was looking for v4 and our DUO setup was set to v3, so this replacement never happened and resulted in a malformed URL leading to 403 responses from the DUO server.

Signed-off-by: smxjrz <Smelecs@gmail.com>